Move Draco DLL loading into io.com
